rename folders to better describe the contents
