+++ apps/api/internal/repository/user_repository.go
@@ -3,15 +3,15 @@ package repository
 import (
 	"context"
 
-	"github.com/Neel-shetty/clarity/internal/domain"
+	"github.com/Neel-shetty/clarity/internal/model"
 	"github.com/google/uuid"
 	"gorm.io/gorm"
 )
 
 type UserRepository interface {
-	CreateUser(ctx context.Context, user *domain.User) error
-	CheckUserExist(ctx context.Context, email string) (*domain.User, error)
-	FindByID(ctx context.Context, userID uuid.UUID) (*domain.User, error)
+	CreateUser(ctx context.Context, user *model.User) error
+	CheckUserExist(ctx context.Context, email string) (*model.User, error)
+	FindByID(ctx context.Context, userID uuid.UUID) (*model.User, error)
 }
 
 type userRepository struct {
@@ -22,8 +22,8 @@ func NewUserRepository(db *gorm.DB) UserRepository {
 	return &userRepository{db}
 }
 
-func (r *userRepository) CheckUserExist(ctx context.Context, email string) (*domain.User, error) {
-	var user domain.User
+func (r *userRepository) CheckUserExist(ctx context.Context, email string) (*model.User, error) {
+	var user model.User
 	err := r.db.WithContext(ctx).Where("email=?", email).First(&user).Error
 	if err != nil {
 		return nil, err
@@ -31,13 +31,13 @@ func (r *userRepository) CheckUserExist(ctx context.Context, email string) (*dom
 	return &user, nil
 }
 
-func (r *userRepository) CreateUser(ctx context.Context, user *domain.User) error {
+func (r *userRepository) CreateUser(ctx context.Context, user *model.User) error {
 	res := r.db.WithContext(ctx).Create(user)
 	return res.Error
 }
 
-func (r *userRepository) FindByID(ctx context.Context, userID uuid.UUID) (*domain.User, error) {
-	var user domain.User
+func (r *userRepository) FindByID(ctx context.Context, userID uuid.UUID) (*model.User, error) {
+	var user model.User
 	err := r.db.WithContext(ctx).Where("id=?", userID).First(&user).Error
 	if err != nil {
 		return nil, err
