+++ apps/api/internal/service/user_service.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"errors"
 
-	"github.com/Neel-shetty/clarity/internal/domain"
+	"github.com/Neel-shetty/clarity/internal/model"
 	"github.com/Neel-shetty/clarity/internal/repository"
 	"github.com/google/uuid"
 	"golang.org/x/crypto/bcrypt"
@@ -12,9 +12,9 @@ import (
 )
 
 type UserService interface {
-	Signup(ctx context.Context, signup *domain.SignUp) error
-	Login(ctx context.Context, login *domain.Login) (*domain.User, error)
-	GetProfile(ctx context.Context, userID uuid.UUID) (*domain.User, error)
+	Signup(ctx context.Context, signup *model.SignUp) error
+	Login(ctx context.Context, login *model.Login) (*model.User, error)
+	GetProfile(ctx context.Context, userID uuid.UUID) (*model.User, error)
 }
 type userService struct {
 	repo repository.UserRepository
@@ -24,20 +24,20 @@ func NewUserService(repo repository.UserRepository) UserService {
 	return &userService{repo}
 }
 
-func (s *userService) Signup(ctx context.Context, signup *domain.SignUp) error {
+func (s *userService) Signup(ctx context.Context, signup *model.SignUp) error {
 	if signup.Password != signup.CheckPassword {
 		return errors.New("passwords do not match")
 	}
 
 	exists, _ := s.repo.CheckUserExist(ctx, signup.Email)
 	if exists != nil {
-		return errors.New("User with this email already exists")
+		return errors.New("user with this email already exists")
 	}
 	hashPass, err := bcrypt.GenerateFromPassword([]byte(signup.Password), bcrypt.DefaultCost)
 	if err != nil {
 		return err
 	}
-	user := &domain.User{
+	user := &model.User{
 		Name:         signup.Name,
 		Email:        signup.Email,
 		PasswordHash: string(hashPass),
@@ -50,7 +50,7 @@ func (s *userService) Signup(ctx context.Context, signup *domain.SignUp) error {
 	return nil
 }
 
-func (s *userService) Login(ctx context.Context, login *domain.Login) (*domain.User, error) {
+func (s *userService) Login(ctx context.Context, login *model.Login) (*model.User, error) {
 	user, err := s.repo.CheckUserExist(ctx, login.Email)
 	if err != nil {
 		if errors.Is(err, gorm.ErrRecordNotFound) {
@@ -65,7 +65,7 @@ func (s *userService) Login(ctx context.Context, login *domain.Login) (*domain.U
 	}
 	return user, nil
 }
-func (s *userService) GetProfile(ctx context.Context, userID uuid.UUID) (*domain.User, error) {
+func (s *userService) GetProfile(ctx context.Context, userID uuid.UUID) (*model.User, error) {
 	user, err := s.repo.FindByID(ctx, userID)
 
 	if err != nil {
